moved each animal class into newly made corresponding modules, changed name of animal.py to index.py, and also imported all of our classes into index.py
