+++ index.py
@@ -1,0 +1,39 @@
+# import the python datetime module to help us create a timestamp
+from datetime import date
+from slithering import (Boa, Python, Splitjaw, Sunbeam, Viper)
+from swimming import (Catfish, Koi, Newt, Toad, Turtle)
+from walking import (Alpaca, Goat, Horse, Llama, Sheep)
+
+miss_fuzz = Llama("Miss Fuzz", "domestic llama")
+
+peaches = Goat("Peaches", "domestic goat")
+
+patches = Sheep("Goat", "domestic sheep")
+
+alfonz = Alpaca("Alfonz", "domestic alpaca")
+
+kaptain = Horse("Kaptain", "Mustang")
+
+pybert = Python("Pybert", "Ball Python")
+
+bob = Boa("Bob", "Long Tail Boa")
+
+val = Viper("Val", "Cottonmouth")
+
+shorty = Sunbeam("Shorty", "Shining sunbeam")
+
+viktor = Splitjaw("Viktor", "smooth-scaled splitjaw")
+
+john = Turtle("John", "Painted turtle")
+
+neal = Newt("Neal", "Eastern newt")
+
+kyle = Koi("Kyle", "Bekko")
+
+kat = Catfish("Kat", "Blue catfish")
+
+jiraya = Toad("Jiraya", "scroll toad")
+
+
+print(miss_fuzz.date_added)
+print(jiraya.name)
