+++ swimming/__init__.py
@@ -1,0 +1,5 @@
+from .turtle import Turtle
+from .catfish import Catfish
+from .koi import Koi
+from .newt import Newt
+from .toad import Toad
