+++ swimming/newt.py
@@ -1,0 +1,8 @@
+from datetime import date
+
+class Newt:
+    def __init__(self, name, species) -> None:
+        self.name = name
+        self.species = species
+        self.date_added = date.today()
+        self.swimming = True
