+++ walking/__init__.py
@@ -1,0 +1,5 @@
+from .llama import Llama
+from .alpaca import Alpaca
+from .goat import Goat
+from .horse import Horse
+from .sheep import Sheep
