+++ walking/alpaca.py
@@ -1,0 +1,9 @@
+from datetime import date
+
+class Alpaca:
+    def __init__(self, name, species) -> None:
+        self.name = name
+        self.species = species
+        self.date_added = date.today()
+        self.walking = True
+
